+++ src/main/java/com/example/sketchTalk/model/Emotion.java
@@ -1,4 +1,4 @@
-package com.example.sketchTalk;
+package com.example.sketchTalk.model;
 
 public enum Emotion {
     Happy, Sad, Angry, Neutral
