+++ src/main/java/com/example/sketchTalk/repository/DiaryRepository.java
@@ -1,6 +1,6 @@
-package com.example.sketchTalk.dialog.repository;
+package com.example.sketchTalk.repository;
 
-import com.example.sketchTalk.model.Diary;
+import com.example.sketchTalk.model.entity.Diary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DiaryRepository extends JpaRepository<Diary, Long> {
